+++ screens/index.js
@@ -1,0 +1,2 @@
+export { default as HomeScreen } from "./home";
+export { default as DetailsScreen } from "./details";
